requirements.txt
Django==5.2
graphene-django==3.2.3
pytest==8.3.5
pytest-django==4.11.1
django-graphql-jwt==0.4.0
